requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appwrite==5.0.2
av==11.0.0
bcrypt==4.1.2
bidict==0.23.1
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
cryptography==42.0.5
ctranslate2==4.1.0
demucs==4.0.1
distro==1.9.0
dora_search==0.1.12
einops==0.7.0
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SSE==1.0.0
flatbuffers==24.3.7
fsspec==2024.2.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
julius==0.2.7
lameenc==1.7.0
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
multiple_datasets @ git+https://github.com/bayartsogt-ya/whisper-multiple-hf-datasets.git@73d9f012cb19f4e0cfe43820b12b6a9da8eb8ea1
mutagen==1.47.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.17.1
openai==1.35.7
openai-whisper==20231117
openpyxl==3.1.2
openunmix==1.3.0
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.32.3
retrying==1.3.4
rsa==4.9
safetensors==0.4.2
scipy==1.13.1
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sshtunnel==0.4.0
stable-ts==2.16.0
submitit==1.5.1
sympy==1.12
tenacity==8.3.0
termcolor==2.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.38.2
treetable==0.2.5
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yt-dlp==2024.7.25
